requirements.txt
flask
Flask-SQLAlchemy
werkzeug
click
greenlet
Markupsafe
python-decouple
Flask-Migrate
psycopg2
